patcher/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
keystone-engine==0.9.2
requests==2.31.0
soupsieve==2.5
types-beautifulsoup4==4.12.0.20240106
types-html5lib==1.1.11.20240106
types-requests==2.31.0.20240106
urllib3==2.1.0
